Code updates. Added separate files for server functions
